refractor: FE OverviewPage -> Move Email logic to EmailClient & Some cardlogic to cardcontainer
